src/requirements.txt
feedgen==1.0.0
feedparser==6.0.11
Flask==2.3.2
functions-framework==3.4.0
google-cloud-logging==3.5.0
google-cloud-storage==2.15.0 
requests==2.31.0
